fix(test): shared multi-line SSE parser for all embedded probes; cold bring-up headroom
Consolidates the Streamable-HTTP SSE parsing (multi-line data: framing
that a 78-tool tools/list response produces) into one verified helper
(tests/src/e2e/utilities/streamable_http.py) used by the embedded smoke
test, the HAOS embedded probe, and the conftest embedded readiness gate,
replacing three divergent copies. Bumps the e2e-validation lane timeout
15 -> 20 minutes: the smoke test's cold bring-up (runtime pip install of
the server tree in its dedicated container) is unique first-install
coverage and now actually runs after the uv-build fix.
